requirements.txt
termcolor
conan
mock
webtest
